build(webpack/build): enable runtime template compiler for Vue
Ideally we'd use precompiled templates with `vue-template-loader` for
bundle size savings as well as a minor performance boost.  This has
tradeoffs though.  The import syntax is more magical and it requires a
few unintuitive changes to Typescript to support this.  It doesn't seem
to work correctly even after jumping through these hoops, so it may be
worth doing this later.